XML loaders, more symbols
Added a few basefile XMLs (FC0N, FC0U, FD0F and FD0G) and changed name
on FD0M+FD0I to match a standard. Also disabled the question if one
wants to load the XMLs + disabled the creation of a new XML file in the
repo for each new filename.