Temporary fix
This is not an actual fix since it reverts to the old behaviour where the user have to manually update a table viewer (or use the interval refresh-thingy in settings) but at least the viewer shows. (it does however read data from sram once when opening it)

I noticed the problem does not affect "gridViewSymbols_DoubleClick" even tho it's making the same calls. Might be a hint for an actual fix in the future.