mark DisplAdap. as calib.
In reality, it's much more complex than this.
What you'd want to do is determine size and offset of the symbol data area in flash and use that in conjunction with with sram offset. If it falls within this area, it should be applied, if not, fall back to old behaviour.
This one is 100% known to always be in flash and sram but very few symbols behave this way.

I made a new branch on my end since actually getting all of that to work will take a while and it _WILL_ break stuff.